@@ -1,5 +1,5 @@
 {
-  "name": "graphql-ws-aws-iot-ws-transport-example",
+  "name": "graphql-aws-iot-transport-example",
   "version": "0.0.0",
   "license": "MIT",
   "scripts": {
@@ -9,44 +9,32 @@
     "test": "karma start ./karma.conf.js",
     "lint": "ng lint",
     "e2e": "protractor ./protractor.conf.js",
-    "schema-generate": "apollo-codegen generate src/graphql/**/*.graphql --schema ./schema.json --target typescript --output src/graphql/schema.ts",
+    "schema:generate": "apollo-codegen generate src/graphql/**/*.graphql --schema ./schema.json --target typescript --output src/graphql/schema.ts",
     "pree2e": "webdriver-manager update --standalone false --gecko false --quiet"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^4.0.0",
-    "@angular/common": "^4.0.0",
-    "@angular/compiler": "^4.0.0",
-    "@angular/core": "^4.0.0",
-    "@angular/forms": "^4.0.0",
-    "@angular/http": "^4.0.0",
-    "@angular/platform-browser": "^4.0.0",
-    "@angular/platform-browser-dynamic": "^4.0.0",
-    "@angular/router": "^4.0.0",
+    "@angular/animations": "^4.3.6",
+    "@angular/cdk": "^2.0.0-beta.10",
+    "@angular/common": "^4.3.0",
+    "@angular/compiler": "^4.3.0",
+    "@angular/core": "^4.3.0",
+    "@angular/forms": "^4.3.0",
+    "@angular/http": "^4.3.0",
+    "@angular/material": "^2.0.0-beta.10",
+    "@angular/platform-browser": "^4.3.0",
+    "@angular/platform-browser-dynamic": "^4.3.0",
+    "@angular/router": "^4.3.0",
     "@types/isomorphic-fetch": "0.0.34",
     "apollo-angular": "0.13.0",
     "apollo-client": "1.9.1",
     "apollo-codegen": "^0.16.5",
-    "aws-sdk": "^2.94.0",
-    "aws4": "^1.6.0",
-    "core-js": "^2.4.1",
-    "domain": "0.0.1",
-    "eventemitter3": "^2.0.3",
-    "graphql": "^0.11.1",
-    "graphql-aws-iot-ws-transport": "file:../qraphl-aws-iot-ws-transport",
-    "graphql-subscriptions": "^0.4.4",
+    "graphql": "0.11.2",
+    "graphql-aws-iot-ws-client": "git+https://github.com/ioxe/graphql-aws-iot-ws-client.git",
     "graphql-tag": "^2.4.2",
-    "lodash": "^4.17.4",
-    "lodash.isobject": "^3.0.2",
-    "lodash.isstring": "^4.0.1",
+    "hammerjs": "^2.0.8",
     "moment": "^2.18.1",
-    "ng2-mqtt": "^0.1.2",
-    "querystring": "^0.2.0",
     "rxjs": "^5.4.1",
-    "stream": "0.0.2",
-    "subscriptions-transport-ws": "^0.8.2",
-    "symbol-observable": "^1.0.4",
-    "url": "^0.11.0",
     "uuid": "^3.1.0",
     "zone.js": "^0.8.14"
   },
@@ -59,6 +47,7 @@
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "^6.0.85",
     "autoprefixer": "^6.5.3",
+    "aws-sdk": "^2.106.0",
     "codelyzer": "~3.0.1",
     "css-loader": "^0.28.1",
     "cssnano": "^3.10.0",
